Refactor/tech-debt: worldgen quality pass — typed handles, split utils, CI, constants
Phases 1–4 of the improvement audit for the worldgen ML pipeline:

**Phase 1 — Safety & correctness**
- Replace `assert` statements in pano_seg/pano_inpaint with `ValueError` (assert
  is stripped in optimized mode)
- Narrow all bare `except Exception/except:` blocks to `except ImportError:` for
  nunchaku/LoRA import guards
- Remove hardcoded `pano_inpainted_image.save(...)` that wrote to CWD on every call
- Fix `[Auroch Syna]` prefix in lora_utils (was "[WorldGen]")

**Phase 2 — Device & dtype centralisation**
- Add `device_utils.py` with `resolve_device()` / `default_dtype()`; wire through
  worldgen.py, pano_gen.py, model_client.py — eliminates the scattered
  `device == "cuda"` string comparisons and the hardcoded `mps` default
- Fix `device.type == "cuda"` comparisons (was comparing torch.device to string)

**Phase 3 — Typed depth result + constants**
- Introduce `DepthPrediction` dataclass in pano_depth.py; replace all
  `predictions["key"]` dict access in worldgen.py, pano_sharp.py, general_utils.py
- Add `constants.py` (MAX_DEPTH_DISTANCE, SH_C0, SHARP_INTERNAL_SHAPE,
  DEFAULT_CUBEMAP_FACE_SIZE); replace magic numbers across pano_sharp, splat_utils
- Merge duplicate `pred_depth`/`pred_pano_depth` functions; keep alias for compat

**Phase 4 — Architecture**
- Move FLUX pipelines from `models/` → `diffusers_ext/`; old paths are one-line
  re-export shims
- Split monolithic `general_utils.py` into `image_utils.py` (pano/cube/resize/map)
  and `geometry_utils.py` (rays/dot/depth_match/mesh); general_utils becomes a
  backward-compat shim
- Add `model_handle.py` with typed handles (DepthHandle, PanoGenHandle, SegHandle,
  InpaintHandle, SharpHandle); ModelClient.build_* returns handles; WorldGen calls
  handle methods instead of bare module functions — fixes two latent bugs where
  `seg_pano_fg` and `inpaint_image` were called without being imported
- Move nunchaku to `[project.optional-dependencies.low-vram]` with Linux/x86_64
  platform marker

**CI & tests**
- Add `.github/workflows/ci.yml` (lint, typecheck, unit-tests, smoke-import)
- Add `requirements-lock.txt` with git-pinned SHAs for viser and utils3d
- Add CPU-only pytest suite: test_general_utils, test_splat_utils, test_pano_seg,
  test_pano_inpaint, test_lora_utils, test_depth_prediction
