+++ src/auroch_syna/worldgen/utils/__init__.py
@@ -1,20 +1,22 @@
-from .general_utils import (
+from .image_utils import (
     pano_to_cube,
     cube_to_pano,
     resize_img,
     resize_img_and_rays,
-    pano_unit_rays,
-    batch_nearest_dot,
     fill_mask_from_contour,
     map_image_to_pano,
-    depth_match
 )
-
+from .geometry_utils import (
+    pano_unit_rays,
+    batch_nearest_dot,
+    depth_match,
+    convert_rgbd2mesh_panorama,
+)
 from .splat_utils import (
     SplatFile,
     convert_rgbd_to_gs,
     mask_splat,
-    merge_splats
+    merge_splats,
 )
 
 __all__ = [
@@ -27,9 +29,9 @@ __all__ = [
     "fill_mask_from_contour",
     "map_image_to_pano",
     "depth_match",
-    
+    "convert_rgbd2mesh_panorama",
     "SplatFile",
     "convert_rgbd_to_gs",
     "mask_splat",
-    "merge_splats"
+    "merge_splats",
 ]
